+++ bespokeasm/assembler/model/operand/types/numeric_expression.py
@@ -6,21 +6,14 @@ from bespokeasm.assembler.model.operand import OperandWithArgument, OperandType,
 
 
 class NumericExpressionOperand(OperandWithArgument):
-    def __init__(self, operand_id: str, arg_config_dict: dict, default_endian: str):
-        super().__init__(operand_id, arg_config_dict, default_endian)
-        # validate config
-        if 'argument' not in self._config:
-            sys.exit(f'ERROR: configuration for numeric operand {self} does not have an arument configuration')
+    def __init__(self, operand_id: str, arg_config_dict: dict, default_endian: str, require_arg: bool = True) -> None:
+        super().__init__(operand_id, arg_config_dict, default_endian, require_arg)
 
     def __str__(self):
         return f'NumericExpressionOperand<{self.id}>'
     @property
     def type(self) -> OperandType:
         return OperandType.NUMERIC
-
-    @property
-    def has_argument(self) -> bool:
-        return True
 
     @property
     def match_pattern(self) -> str: